WIP: asm goto: x86: convert some __always_inline functions to macros
The use of the "i" constraint modifier in inline assembly with operands
of inline functions is problematic, as it is code that is only valid in
GCC at -O2 (but results in a compilation error at -O0).

Example: https://godbolt.org/z/ttM5Eg

"i":
An immediate integer operand (one with constant value) is allowed. This
includes symbolic constants whose values will be known only at assembly
time or later.
https://gcc.gnu.org/onlinedocs/gcc/Simple-Constraints.html#Simple-Constraints

The operands of a static inline function are NOT immediate integer
operands, unless they are inlined into a callsite whose arguments are
immediate integer operands.

That this even works, but only at -O2, takes advantage of the
implementation detail of GCC, in which semantic analysis is performed
somewhat during optimizations (at least after inlining).

The correct way to rely on the inline asm inputs to be compatible with
"i" is to convert the inline functions into macros; as the C
preprocessor is guaranteed to perform the inlining before any semantic
analysis is run. Further, to continue to be used as expressions, they
are converted to GNU statement expressions.  Finally, in order to
prevent multiple label definition errors (labels in C are function
scoped) in functions that call these new macros multiple times, we take
advantage of local labels, a GNU C extension.